QUIPU-287 delete unused interface
