+++ src/modules/new-liquidity/interfaces/index.ts
@@ -3,4 +3,4 @@ export * from './liquidity-item.interface';
 export * from './liquidity-token-info.interface';
 export * from './new-liquidity-lables.interface';
 export * from './opportunities.interface';
-export * from './hot-pool.interface';
+export * from './pool-type-enum';
